Refactor signup form validation
[add]
core error responses
