+++ api/app/auth/routes.py
@@ -1,20 +1,19 @@
 from flask import Blueprint, jsonify, request
+from http import HTTPStatus
 from app import db
 from app.users.models import User
 from app.auth.forms import SignUpForm
+from core.http.error_responses import UnprocessableEntityResponse
 
 auth = Blueprint("auth", __name__, url_prefix="/api/auth")
 
+    
 @auth.route("/signup", methods=["POST"])
 def sign_up():
     form = SignUpForm()
     if not form.validate():
-        response = jsonify({
-            "error": "unprocessable entity",
-            "fields": dict(form.errors.items())
-        })
-        response.status_code = 422
-        return response
+        return UnprocessableEntityResponse(fields=dict(form.errors.items())).jsonify()
+
     username = request.form.get("username")
     password = request.form.get("password")
     user = User(username=username, password=password)
